requirements.txt
streamlit

pathlib
python-dotenv
youtube_transcript_api
google-generativeai